Add the 'prep' phase
This phase 'prepare' a directory tree suitable for calling odocmkgen on
it.
This abstracts specific details like the notion of "package" and the
layout used to represent them from the main tool.
The goal is to define an input for odocmkgen that is both simpler and
exposes every features of Odoc (especially parent-child).